ingestion/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
